update budgets settings
